Auth page styling and organization
