+++ client/src/app/App.tsx
@@ -1,7 +1,16 @@
 import * as React from 'react';
+import style from './App.st.css';
 
-export class App extends React.Component {
-  render(): React.ReactNode {
-    return <div>Hello, World!</div>;
-  }
-}
+import { Logo } from 'shared/Logo';
+
+export const App: React.SFC = props => {
+  React.useEffect(() => {
+    document.title = 'myx';
+  });
+
+  return (
+    <div {...style('root', {}, props)}>
+      <Logo />
+    </div>
+  );
+};
