+++ client/src/shared/Logo.st.css
@@ -1,13 +1,19 @@
+@namespace 'Logo';
+:import {
+  -st-from: 'project.st.css';
+  -st-named: brandblue, gray3;
+}
+
 .root {
   font-weight: 900;
-  color: #8d8d9a;
+  color: value(gray3);
 }
 
 .m {
 }
 
 .y {
-  color: #1da1f2;
+  color: value(brandblue);
 }
 
 .x {
