--- client/src/signup/SignUp.tsx
@@ -1,8 +1,0 @@
-import * as React from 'react';
-
-export const SignUp: React.SFC = () => {
-  React.useEffect(() => {
-    document.title = 'Myx - Sign Up';
-  });
-  return <div>SignUp</div>;
-};
